Import HederaParamsFactory directly
