+++ src/helpers/HederaParamsFactory.ts
@@ -1,5 +1,20 @@
 import { AccountId, RequestType, Transaction } from "@hashgraph/sdk";
-import { HederaSignAndSendTransactionParams } from "./types";
+
+import { TypedRequestParams } from "./types";
+
+/**
+ * `transaction.type` should be an instance of `RequestType` from `@hashgraph/sdk`.
+ * e.g. RequestType.CryptoTransfer.toString()
+ */
+export type HederaSignAndSendTransactionParams = {
+  transaction: {
+    type: string;
+    bytes: Uint8Array;
+  };
+};
+
+export type HederaSessionRequestParams =
+  TypedRequestParams<HederaSignAndSendTransactionParams>;
 
 export class HederaParamsFactory {
   public static buildSignAndSendTransactionPayload(
