+++ src/helpers/index.ts
@@ -4,4 +4,3 @@ export * from "./eip1271";
 export * from "./tx";
 export * from "./types";
 export * from "./utilities";
-export * from "./hedera";
